refactor(tools): move registry.rs's test module out of the file it tests
registry.rs was 3791 lines, more than half of it an inline `mod tests`.
Four sibling test modules already lived in registry/ (fence, gate_batch,
file_change, private_state); this one was the odd one out.

registry.rs: 3791 -> 2015 lines, and the baseline ceiling ratchets down
to match, which is what scripts/check-file-size.sh asks for after a
split.

The tests land as registry/tests.rs plus registry/tests/{schema_gate,
touch,chain}.rs rather than one file, because a new 1787-line file is a
hard block under that same guard — correctly, since moving a navigation
problem is not solving it. The split follows the section banners the
module already had.

Two structural notes, both discovered by the compiler:

  * `mod tests;` carries NO `#[path]`, unlike the four siblings. An
    explicit path on a parent makes rustc resolve that module's OWN
    children against registry/ instead of registry/tests/, so the
    submodules were not found. Plain `mod tests;` resolves both levels.
  * `telemetry_fixture`/`exec_ok` moved up to tests.rs: `chain` uses
    them too, and sibling modules cannot see each other's helpers.
    `seeded_snapshot` stayed put for the same class of reason — it is
    `pub(super)` and registry::gate_batch_tests reaches it as
    `super::tests::seeded_snapshot`.

No test was added, removed or edited: 557 pass before and after.
