+++ stella-tools/src/registry.rs
@@ -2006,1786 +2006,10 @@ mod gate_batch_tests;
 #[path = "registry/file_change_tests.rs"]
 mod file_change_tests;
 
+// No `#[path]` here, unlike the four single-file siblings above: this one has
+// child modules of its own (`registry/tests/*.rs`), and an explicit `#[path]`
+// on the parent makes rustc resolve those children against `registry/` rather
+// than `registry/tests/`. Plain `mod tests;` resolves to `registry/tests.rs`
+// and its children to `registry/tests/`, which is what the layout wants.
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    use crate::issues::IssueBackend;
-
-    /// A registry rooted in a fresh empty tempdir. Rooting tests at a shared
-    /// path like `/tmp` is not hermetic: a stray `.stella/private/codegraph.db` left
-    /// there by a real session conditionally registers `graph_query` and
-    /// skews every tool-set assertion. The `TempDir` is returned so the root
-    /// outlives the registry.
-    fn bare_registry(issue_backend: Option<IssueBackend>) -> (tempfile::TempDir, ToolRegistry) {
-        let root = tempfile::tempdir().unwrap();
-        let reg = ToolRegistry::with_issue_backend(root.path().to_path_buf(), issue_backend);
-        (root, reg)
-    }
-
-    /// `names` with `web_search` folded in when the *host running the tests*
-    /// happens to export a BYOK search key. `web_search` is the one remaining
-    /// environment-dependent row in the default surface — the other three web
-    /// tools need no key — so an exact-set pin has to account for it rather
-    /// than fail on a developer's machine.
-    fn with_ambient_search(mut names: Vec<&'static str>) -> Vec<&'static str> {
-        if crate::web::detect_search_backend().is_some() {
-            names.push("web_search");
-            names.sort_unstable();
-        }
-        names
-    }
-
-    /// A coverage hint must fire only on whole path tokens — a substring hit
-    /// (`lib.rs` inside `mylib.rs`) would permanently burn the map's
-    /// once-per-session hint on a file it doesn't cover.
-    #[test]
-    fn mentions_path_requires_token_boundaries() {
-        assert!(mentions_path("src/lib.rs:12: pub fn x()", "src/lib.rs"));
-        assert!(mentions_path("lib.rs", "lib.rs"));
-        // An absolute spelling of a covered relative path still matches.
-        assert!(mentions_path("/tmp/ws/src/lib.rs:3: fn y()", "src/lib.rs"));
-        assert!(mentions_path(
-            "see `core/driver.rs` for the loop",
-            "core/driver.rs"
-        ));
-        assert!(!mentions_path("mylib.rs:3: struct Y", "lib.rs"));
-        assert!(!mentions_path("graphlib.rs/index.js", "lib.rs"));
-        assert!(!mentions_path("src/lib.rs.bak", "src/lib.rs"));
-        // A miss on one occurrence must not mask a clean later occurrence.
-        assert!(mentions_path("mylib.rs and also lib.rs here", "lib.rs"));
-        assert!(!mentions_path("anything", ""));
-    }
-
-    #[tokio::test]
-    async fn unknown_tool_returns_error_not_panic() {
-        let (_root, reg) = bare_registry(None);
-        let result = reg.execute("nonexistent", &Value::Null).await;
-        assert!(result.is_error());
-    }
-
-    /// The measured defect: over a 20-task Terminal-Bench run, 757 of 1,063
-    /// tool calls were `bash` and the ledger recorded none of them, because
-    /// `classify_file_op` can only read a tool's *input* and a shell command
-    /// is an opaque string. With the diff probe dark on a non-git task
-    /// directory, `file_change_events` is the only channel left that can
-    /// prove the tree changed — so a blind ledger there is a blind ladder.
-    #[tokio::test]
-    async fn a_file_written_by_the_shell_reaches_the_ledger() {
-        let (root, reg) = bare_registry(None);
-        assert_eq!(reg.mutations_recorded(), 0, "clean before");
-
-        reg.begin_workspace_probe();
-        let out = reg
-            .execute(
-                "bash",
-                &serde_json::json!({"command": "printf 'a\\nb\\n' > made_by_shell.txt"}),
-            )
-            .await;
-        assert!(!out.is_error(), "{out:?}");
-        assert!(root.path().join("made_by_shell.txt").exists());
-        assert_eq!(
-            reg.mutations_recorded(),
-            0,
-            "nothing is attributed until the turn settles"
-        );
-        reg.settle_workspace_probe();
-
-        let touched = reg.files_touched();
-        assert!(
-            touched
-                .iter()
-                .any(|(path, ops)| path == "made_by_shell.txt" && ops.contains('C')),
-            "shell-created file missing from the ledger: {touched:?}"
-        );
-        // The count the verification ladder reads.
-        assert!(reg.mutations_recorded() >= 1, "no mutation recorded");
-    }
-
-    /// A shell command that fails can still have written something, so
-    /// attribution must not hang off the exit code.
-    #[tokio::test]
-    async fn a_failing_shell_command_still_reports_what_it_wrote() {
-        let (root, reg) = bare_registry(None);
-
-        reg.begin_workspace_probe();
-        let out = reg
-            .execute(
-                "bash",
-                &serde_json::json!({"command": "echo partial > half_done.txt; exit 3"}),
-            )
-            .await;
-        // Whatever the tool reports, the tree changed and the ledger says so.
-        let _ = out;
-        reg.settle_workspace_probe();
-        assert!(root.path().join("half_done.txt").exists());
-        assert!(
-            reg.files_touched()
-                .iter()
-                .any(|(path, _)| path == "half_done.txt"),
-            "a non-zero exit erased the attribution"
-        );
-    }
-
-    /// Reads must not be inflated into mutations: a shell command that only
-    /// looks at the tree leaves the ledger where it found it.
-    #[tokio::test]
-    async fn a_read_only_shell_command_records_no_mutation() {
-        let (root, reg) = bare_registry(None);
-        std::fs::write(root.path().join("existing.txt"), "unchanged\n").unwrap();
-
-        reg.begin_workspace_probe();
-        let out = reg
-            .execute("bash", &serde_json::json!({"command": "cat existing.txt"}))
-            .await;
-        assert!(!out.is_error(), "{out:?}");
-        reg.settle_workspace_probe();
-        assert_eq!(
-            reg.mutations_recorded(),
-            0,
-            "inspecting the tree is not changing it: {:?}",
-            reg.files_touched()
-        );
-    }
-
-    /// The per-call probe is a fallback, not dead code. A host that never
-    /// brackets a turn — every embedder predating the pair, and every test
-    /// double — must keep exactly the attribution it already had, or gating
-    /// the fast path would have silently deleted the slow one.
-    #[tokio::test]
-    async fn an_unbracketed_session_still_attributes_the_shell_per_call() {
-        let (root, reg) = bare_registry(None);
-        // Deliberately no begin_workspace_probe(): this host does not bracket.
-        let out = reg
-            .execute(
-                "bash",
-                &serde_json::json!({"command": "echo hi > unbracketed.txt"}),
-            )
-            .await;
-        assert!(!out.is_error(), "{out:?}");
-        assert!(root.path().join("unbracketed.txt").exists());
-        assert!(
-            reg.files_touched()
-                .iter()
-                .any(|(path, _)| path == "unbracketed.txt"),
-            "gating the turn probe cost an unbracketed host its attribution"
-        );
-    }
-
-    /// The two granularities are mutually exclusive, and this is the property
-    /// that says so. Running both walks the tree twice per shell call *and*
-    /// twice per turn to learn one fact — which on a 900s trial with 757 shell
-    /// calls costs more than the trial it exists to observe.
-    #[tokio::test]
-    async fn a_bracketed_session_attributes_each_shell_write_exactly_once() {
-        let (_root, reg) = bare_registry(None);
-
-        reg.begin_workspace_probe();
-        for i in 0..3 {
-            let out = reg
-                .execute(
-                    "bash",
-                    &serde_json::json!({"command": format!("echo {i} > f{i}.txt")}),
-                )
-                .await;
-            assert!(!out.is_error(), "{out:?}");
-        }
-        assert_eq!(
-            reg.mutations_recorded(),
-            0,
-            "the per-call probe fired inside a bracket: {:?}",
-            reg.files_touched()
-        );
-
-        reg.settle_workspace_probe();
-        assert_eq!(
-            reg.mutations_recorded(),
-            3,
-            "expected one mutation per created file, no double-attribution: {:?}",
-            reg.files_touched()
-        );
-    }
-
-    /// The probe holds a real pre-image for every file it can fit in budget,
-    /// and the counts must come from *that*. Re-reading the file at settle
-    /// time yields its POST-image, which renders every shell rewrite as
-    /// 0 added / 0 removed — the same blindness this module exists to remove,
-    /// one level further down.
-    #[tokio::test]
-    async fn a_shell_rewrite_reports_the_line_counts_it_measured() {
-        let (root, reg) = bare_registry(None);
-        std::fs::write(root.path().join("rewritten.txt"), "one\ntwo\nthree\n").unwrap();
-
-        reg.begin_workspace_probe();
-        let out = reg
-            .execute(
-                "bash",
-                &serde_json::json!({
-                    "command": "printf 'one\\ntwo\\nthree\\nfour\\n' > rewritten.txt"
-                }),
-            )
-            .await;
-        assert!(!out.is_error(), "{out:?}");
-        reg.settle_workspace_probe();
-
-        let telemetry = reg.file_touch_telemetry();
-        let record = telemetry
-            .files_touched
-            .iter()
-            .find(|r| r.path == "rewritten.txt")
-            .expect("the rewritten file never reached the ledger");
-        // One line appended: LCS trims the shared prefix, leaving 1 added.
-        assert_eq!(
-            (record.lines_added, record.lines_removed),
-            (1, 0),
-            "counts came from the post-image, not the captured pre-image"
-        );
-    }
-
-    #[tokio::test]
-    async fn session_reads_flow_into_saved_exploration_manifest() {
-        let (root, reg) = bare_registry(None);
-        std::fs::write(root.path().join("evidence.rs"), "fn seen() {}").unwrap();
-
-        // Read through the registry so the file-touch ledger records it.
-        let read = reg
-            .execute(
-                "read_file",
-                &serde_json::json!({"path": "evidence.rs", "reason": "test"}),
-            )
-            .await;
-        assert!(!read.is_error(), "{read:?}");
-
-        // Save WITHOUT declaring the file — the ledger must supply it.
-        let saved = reg
-            .execute(
-                "save_exploration",
-                &serde_json::json!({
-                    "slice": "auto", "title": "Auto", "summary": "s", "content": "map"
-                }),
-            )
-            .await;
-        match &saved {
-            ToolOutput::Ok { content } => {
-                assert!(content.contains("1 files tracked"), "{content}")
-            }
-            other => panic!("{other:?}"),
-        }
-    }
-
-    /// The ledger key is the registry's channel, never the model's: with
-    /// zero session reads a model-authored `_session_read_files` used to
-    /// pass through untouched and hash fabricated evidence into the
-    /// manifest.
-    #[tokio::test]
-    async fn model_authored_ledger_key_is_dropped_when_the_session_read_nothing() {
-        let (root, reg) = bare_registry(None);
-        std::fs::write(root.path().join("sneaky.rs"), "fn planted() {}").unwrap();
-
-        let saved = reg
-            .execute(
-                "save_exploration",
-                &serde_json::json!({
-                    "slice": "fab", "title": "Fab", "summary": "s", "content": "map",
-                    crate::exploration::LEDGER_FILES_KEY: ["sneaky.rs"]
-                }),
-            )
-            .await;
-        match &saved {
-            ToolOutput::Ok { content } => {
-                assert!(
-                    content.contains("0 files tracked"),
-                    "no session reads means no ledger evidence: {content}"
-                );
-            }
-            other => panic!("{other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn fresh_map_hints_once_on_covering_search_results() {
-        let (root, reg) = bare_registry(None);
-        std::fs::write(root.path().join("covered.rs"), "fn mapped() {}").unwrap();
-        let saved = reg
-            .execute(
-                "save_exploration",
-                &serde_json::json!({
-                    "slice": "zone", "title": "Zone", "summary": "s", "content": "map",
-                    "files": ["covered.rs"]
-                }),
-            )
-            .await;
-        assert!(!saved.is_error(), "{saved:?}");
-        // The author never needs a hint about its own map — simulate a new
-        // session (fresh registry, same workspace) which does.
-        let reg2 = ToolRegistry::with_issue_backend(root.path().to_path_buf(), None);
-
-        let hit = reg2
-            .execute("grep", &serde_json::json!({"pattern": "mapped"}))
-            .await;
-        match &hit {
-            ToolOutput::Ok { content } => {
-                assert!(
-                    content.contains("saved exploration `zone`"),
-                    "first covering search must carry the hint: {content}"
-                );
-            }
-            other => panic!("{other:?}"),
-        }
-
-        // Once per session: the second covering search stays clean.
-        let again = reg2
-            .execute("grep", &serde_json::json!({"pattern": "mapped"}))
-            .await;
-        match &again {
-            ToolOutput::Ok { content } => {
-                assert!(
-                    !content.contains("saved exploration `zone`"),
-                    "hint must not repeat: {content}"
-                );
-            }
-            other => panic!("{other:?}"),
-        }
-
-        // And the author's own registry was seeded as already-hinted.
-        let author_search = reg
-            .execute("grep", &serde_json::json!({"pattern": "mapped"}))
-            .await;
-        match &author_search {
-            ToolOutput::Ok { content } => {
-                assert!(
-                    !content.contains("saved exploration `zone`"),
-                    "author must not be hinted about its own map: {content}"
-                );
-            }
-            other => panic!("{other:?}"),
-        }
-    }
-
-    #[test]
-    fn every_registry_tool_is_reserved_against_custom_shadowing() {
-        // RESERVED_NAMES is now an alias for catalog::ALL_NAMES rather than a
-        // hand-mirrored copy (#450), so declaring a tool reserves it. What
-        // still needs proving is the *other* direction: a tool that registers
-        // but was never declared is shadowable by a custom manifest.
-        let (_root, reg) = bare_registry(Some(IssueBackend::GitHub));
-        for schema in reg.schemas() {
-            assert!(
-                crate::custom::RESERVED_NAMES.contains(&schema.name.as_str()),
-                "built-in tool `{}` is missing from custom::RESERVED_NAMES — \
-                 declare it in stella-tools/src/catalog.rs, or a custom \
-                 manifest could shadow it",
-                schema.name
-            );
-        }
-
-        // Conditionally-registered tools may not show up in a bare registry's
-        // schemas (the media tools need a capable key, `web_search` a search
-        // key), so the registry-driven loop above can't reach them. The
-        // catalog declares them, and the alias reserves them — assert the
-        // conditional half of the table is genuinely covered.
-        for name in crate::catalog::names_where(|a| {
-            a != crate::catalog::Availability::Always && a.is_native()
-        }) {
-            assert!(
-                crate::custom::RESERVED_NAMES.contains(&name),
-                "conditionally-registered tool `{name}` is not reserved"
-            );
-        }
-        // The CLI's session layer is reserved too — it is not in any registry.
-        for name in ["ask_user", "tool_search", "install_skill"] {
-            assert!(crate::custom::RESERVED_NAMES.contains(&name), "{name}");
-        }
-    }
-
-    /// Witness for #450: the advertised set is pinned as an exact **name set**
-    /// against the canonical catalog, not as a count. A magic-number pin
-    /// (`names.len() == 50`) merges to a plausible-but-wrong integer when two
-    /// parallel PRs each bump it off the same base — this fails by name
-    /// instead, naming exactly what was added or dropped.
-    #[test]
-    fn registry_advertises_exactly_the_catalog_tool_set() {
-        let (_root, reg) = bare_registry(Some(IssueBackend::GitHub));
-        let schemas = reg.schemas();
-        let names: Vec<&str> = schemas.iter().map(|s| s.name.as_str()).collect();
-        assert_eq!(
-            names,
-            with_ambient_search(crate::catalog::always_on_with_issues()),
-            "registry disagrees with catalog::CATALOG — register the tool, or \
-             add/remove its line in stella-tools/src/catalog.rs"
-        );
-        // `bash` IS in the default surface. Switching it off is settings work,
-        // and it happens above this registry (crate::policy::ToolPolicy).
-        assert!(names.contains(&"bash"), "{names:?}");
-    }
-
-    /// Witness for #450: a tool that registers but was never declared in the
-    /// catalog is caught by *name*. Simulated here by dropping a name from the
-    /// expectation — the same failure a real unregistered tool produces.
-    #[test]
-    fn an_undeclared_tool_fails_the_catalog_pin_by_name() {
-        let (_root, reg) = bare_registry(Some(IssueBackend::GitHub));
-        let schemas = reg.schemas();
-        let live: Vec<&str> = schemas.iter().map(|s| s.name.as_str()).collect();
-        let mut catalog_missing_one = with_ambient_search(crate::catalog::always_on_with_issues());
-        let dropped = catalog_missing_one.pop().expect("catalog is non-empty");
-        assert_ne!(
-            live, catalog_missing_one,
-            "a catalog missing `{dropped}` must not compare equal to the live \
-             registry — otherwise the pin cannot catch an undeclared tool"
-        );
-        // And the difference is reported as a name, not an arity.
-        assert!(!catalog_missing_one.contains(&dropped));
-        assert!(live.contains(&dropped));
-    }
-
-    // bash and the web family ship ON (the default flip)
-
-    /// **Witness for the default flip.** With no options at all — no
-    /// settings, no opt-in, nothing — `bash` is advertised AND dispatchable.
-    /// This test fails on the old code, where `RegistryOptions::bash`
-    /// defaulted to `false` and the shell was absent until a settings key
-    /// said otherwise.
-    #[tokio::test]
-    async fn bash_is_registered_with_no_options_at_all() {
-        let (_root, reg) = bare_registry(None);
-        assert!(
-            reg.schemas().iter().any(|s| s.name == "bash"),
-            "bash must be advertised with no configuration"
-        );
-        let out = reg
-            .execute(
-                "bash",
-                &serde_json::json!({"command": "echo bash_default_on"}),
-            )
-            .await;
-        match out {
-            ToolOutput::Ok { content } => assert!(content.contains("bash_default_on")),
-            ToolOutput::Error { message } => panic!("default bash must run: {message}"),
-        }
-        // The default options carry no policy at all any more — the only
-        // switch left is the host attestation.
-        assert!(
-            RegistryOptions::default()
-                .media_host_data_isolation
-                .is_none()
-        );
-    }
-
-    /// **Witness for the default flip, web half.** The three key-free web
-    /// tools register with no options; `web_search` still needs a search key,
-    /// so its presence is environment-dependent and not pinned here.
-    #[test]
-    fn the_key_free_web_family_is_registered_with_no_options_at_all() {
-        let (_root, reg) = bare_registry(None);
-        let schemas = reg.schemas();
-        for (expected, read_only) in [
-            ("web_fetch", true),
-            ("web_extract_assets", true),
-            ("web_download", false),
-        ] {
-            let schema = schemas
-                .iter()
-                .find(|s| s.name == expected)
-                .unwrap_or_else(|| panic!("{expected} must register with no configuration"));
-            assert_eq!(schema.read_only, read_only, "{expected}");
-        }
-    }
-
-    /// The schema list is serialized verbatim into the prompt prefix; a
-    /// nondeterministic order (HashMap iteration) breaks byte-level
-    /// prompt-cache prefix matching across processes.
-    #[test]
-    fn schemas_are_sorted_by_name_for_prompt_cache_stability() {
-        let (_root, reg) = bare_registry(Some(IssueBackend::GitHub));
-        let names: Vec<String> = reg.schemas().iter().map(|s| s.name.clone()).collect();
-        let mut sorted = names.clone();
-        sorted.sort();
-        assert_eq!(names, sorted, "schemas must be name-sorted");
-    }
-
-    #[test]
-    fn issue_tools_absent_without_a_configured_backend() {
-        let (_root, reg) = bare_registry(None);
-        let schemas = reg.schemas();
-        let names: Vec<&str> = schemas.iter().map(|s| s.name.as_str()).collect();
-        // The no-backend surface is a *filtered view* of the same canonical
-        // table, not a second count to keep in sync (#450).
-        assert_eq!(
-            names,
-            with_ambient_search(crate::catalog::always_on()),
-            "the no-backend registry must be exactly the always-on catalog"
-        );
-        for absent in crate::catalog::names_where(|a| a == crate::catalog::Availability::Issue) {
-            assert!(!names.contains(&absent), "{absent} must be absent");
-        }
-    }
-
-    #[test]
-    fn video_tools_register_only_with_a_video_capable_media_backend() {
-        // A provider that satisfies the port but is never called — tool
-        // registration is what's under test here, not generation.
-        struct NullMedia;
-        #[async_trait]
-        impl stella_media::MediaProvider for NullMedia {
-            fn id(&self) -> &str {
-                "null"
-            }
-            fn capabilities(&self) -> stella_media::MediaCapabilities {
-                stella_media::MediaCapabilities::default()
-            }
-            async fn generate_image(
-                &self,
-                _req: stella_media::ImageRequest,
-            ) -> Result<stella_media::MediaArtifact, stella_media::MediaError> {
-                Err(stella_media::MediaError::Terminal("not under test".into()))
-            }
-            async fn generate_video(
-                &self,
-                _req: stella_media::VideoRequest,
-            ) -> Result<stella_media::MediaJob, stella_media::MediaError> {
-                Err(stella_media::MediaError::Terminal("not under test".into()))
-            }
-            async fn poll_video(
-                &self,
-                _job: &stella_media::MediaJob,
-            ) -> Result<stella_media::MediaJobStatus, stella_media::MediaError> {
-                Err(stella_media::MediaError::Terminal("not under test".into()))
-            }
-        }
-        let provider: Arc<dyn stella_media::MediaProvider> = Arc::new(NullMedia);
-        struct FixedIds;
-        impl crate::media::MediaOperationIdSource for FixedIds {
-            fn operation_id(&self) -> crate::media::HostMediaOperation {
-                crate::media::HostMediaOperation {
-                    opaque_id: "op-registry-test".into(),
-                    expires_at: u64::MAX,
-                }
-            }
-        }
-        // A full approving host context: the paid tools register only under
-        // one (#785).
-        let host_options = || RegistryOptions {
-            media_spend_gate: Some(Arc::new(stella_media::DenyMediaSpendGate)),
-            media_operation_ids: Some(Arc::new(FixedIds)),
-            media_operation_journal: Some(Arc::new(
-                stella_media::SqliteMediaOperationJournal::open_in_memory(Default::default())
-                    .unwrap(),
-            )),
-            media_requires_host_approval: true,
-            media_host_data_isolation: Some(crate::media::HostDataIsolation::ProcessFree),
-        };
-        let names = |backend, options: RegistryOptions| {
-            let root = tempfile::tempdir().unwrap();
-            ToolRegistry::with_backends_and_options(
-                root.path().to_path_buf(),
-                None,
-                Some(backend),
-                options,
-            )
-            .schemas()
-            .iter()
-            .map(|s| s.name.clone())
-            .collect::<Vec<_>>()
-        };
-
-        let with_video = names(
-            crate::media::MediaBackend {
-                image: provider.clone(),
-                video: Some(provider.clone()),
-            },
-            host_options(),
-        );
-        for expected in ["generate_image", "generate_video", "poll_video"] {
-            assert!(
-                with_video.contains(&expected.to_string()),
-                "missing {expected}: {with_video:?}"
-            );
-        }
-
-        let image_only = names(
-            crate::media::MediaBackend {
-                image: provider.clone(),
-                video: None,
-            },
-            host_options(),
-        );
-        assert!(image_only.contains(&"generate_image".to_string()));
-        for absent in ["generate_video", "poll_video"] {
-            assert!(
-                !image_only.contains(&absent.to_string()),
-                "{absent} must be absent without a video adapter"
-            );
-        }
-
-        // The #785 ruling: with credentials but NO approving host context,
-        // the paid tools would deny every call — so none of them surface.
-        // The free, client-side generate_svg still does.
-        let deny_only = names(
-            crate::media::MediaBackend {
-                image: provider,
-                video: Some(Arc::new(NullMedia)),
-            },
-            RegistryOptions::default(),
-        );
-        for absent in ["generate_image", "generate_video", "poll_video"] {
-            assert!(
-                !deny_only.contains(&absent.to_string()),
-                "{absent} must not register without a host context"
-            );
-        }
-        assert!(deny_only.contains(&"generate_svg".to_string()));
-    }
-
-    #[tokio::test]
-    async fn graph_query_is_advertised_without_an_index_and_builds_one_on_first_use() {
-        // `graph_query` is advertised from the start, even
-        // in a workspace with no `.stella/private/codegraph.db`. Hiding it
-        // until an index existed hid the very tool that builds one, so the
-        // agent grepped instead. The first call builds the index it needs.
-        let dir = tempfile::tempdir().unwrap();
-        let root = dir.path().to_path_buf();
-        let reg = ToolRegistry::with_issue_backend(root.clone(), None);
-        let has_graph_query =
-            |r: &ToolRegistry| r.schemas().iter().any(|s| s.name == "graph_query");
-
-        assert!(
-            has_graph_query(&reg),
-            "advertised with no index present yet"
-        );
-        assert!(
-            !crate::graph::graph_db_path(&root).exists(),
-            "and no index exists at this point"
-        );
-
-        std::fs::write(root.join("lib.rs"), "pub fn f() {}\n").unwrap();
-        // The first query builds the index and answers — no `stella init`,
-        // no prior `enable_code_graph_if_available`.
-        let out = reg
-            .execute(
-                "graph_query",
-                &serde_json::json!({"op": "definitions", "target": "f"}),
-            )
-            .await;
-        assert!(!out.is_error(), "first-use build must dispatch: {out:?}");
-        assert!(
-            crate::graph::graph_db_path(&root).exists(),
-            "the call built the index on first use"
-        );
-    }
-
-    /// Issue #330 witness: `read_symbol` dispatches through the registry,
-    /// returns exactly the graph-resolved span, lands the same `R`
-    /// file-touch event a `read_file` of that file would, and shares one
-    /// per-file read tally with `read_file`.
-    #[tokio::test]
-    async fn read_symbol_reads_the_exact_span_and_lands_a_file_touch_read_event() {
-        let (root, reg) = bare_registry(None);
-        std::fs::write(
-            root.path().join("lib.rs"),
-            "fn before() {}\n\nfn span_me() {\n    let x = 1;\n}\n",
-        )
-        .unwrap();
-
-        let out = reg
-            .execute(
-                "read_symbol",
-                &serde_json::json!({"name": "span_me", "reason": "witness"}),
-            )
-            .await;
-        match &out {
-            ToolOutput::Ok { content } => {
-                assert!(content.contains("fn span_me (lib.rs:3-5)"), "{content}");
-                assert!(content.contains("3\tfn span_me() {"), "{content}");
-                assert!(!content.contains("before"), "exactly the span: {content}");
-            }
-            other => panic!("read_symbol must dispatch: {other:?}"),
-        }
-
-        // The read landed in the file-touch ledger as an `R` on the resolved
-        // path — audit parity with `read_file`.
-        let touched = reg.files_touched();
-        assert!(
-            touched
-                .iter()
-                .any(|(path, ops)| path == "lib.rs" && ops.contains('R')),
-            "read_symbol must land an R event: {touched:?}"
-        );
-
-        // And the per-file read tally is shared: a `read_file` of the same
-        // file counts as the second read this session.
-        let second = reg
-            .execute("read_file", &serde_json::json!({"path": "lib.rs"}))
-            .await;
-        match &second {
-            ToolOutput::Ok { content } => {
-                assert!(
-                    content.contains("read 2× this session"),
-                    "one tally across both read surfaces: {content}"
-                );
-            }
-            other => panic!("read_file must read: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn read_only_flags_partition_the_registry_correctly() {
-        // The engine parallelizes on this flag — a mutating tool marked
-        // read-only would race writes; a read-only tool marked mutating
-        // just loses concurrency. Pin the partition explicitly.
-        // The expectation is the canonical catalog's flag, so the partition
-        // lives in one place rather than being restated here (#450).
-        let (_root, reg) = bare_registry(Some(IssueBackend::GitHub));
-        for schema in reg.schemas() {
-            let entry = crate::catalog::get(&schema.name).unwrap_or_else(|| {
-                panic!(
-                    "`{}` registers but is not declared in catalog::CATALOG",
-                    schema.name
-                )
-            });
-            assert_eq!(
-                schema.read_only, entry.read_only,
-                "read_only flag wrong for {}",
-                schema.name
-            );
-            assert_eq!(
-                schema.speculation_safe, entry.speculation_safe,
-                "speculation_safe flag wrong for {} — the engine runs \
-                 exactly this partition before a step commits, so a schema \
-                 drifting from the catalog either double-bills a metered \
-                 read or silently stops speculating a pure one (#923)",
-                schema.name
-            );
-        }
-    }
-
-    /// A baseline snapshot with the given tables in the implicit `sql`
-    /// layer, `default` namespace — the shape `stella init` would seed.
-    /// `pub(super)` so the batch-gate witnesses (`gate_batch_tests`) reuse
-    /// the same fixture instead of growing a drifting copy.
-    pub(super) fn seeded_snapshot(tables: &[&str]) -> stella_graph::StorageSnapshot {
-        stella_graph::StorageSnapshot {
-            layers: vec![],
-            relations: tables
-                .iter()
-                .map(|name| stella_graph::storage::RelationEntry {
-                    address: stella_graph::storage::relation_address("sql", "default", name),
-                    layer: "sql".into(),
-                    namespace: "default".into(),
-                    name: name.to_string(),
-                    kind: "table".into(),
-                    fields: vec![stella_graph::storage::FieldEntry {
-                        name: "id".into(),
-                        data_type: Some("INT".into()),
-                        nullable: false,
-                        default_value: None,
-                        constraints: vec!["PRIMARY KEY".into()],
-                        references: None,
-                        intent: None,
-                        line: 1,
-                    }],
-                    enum_values: vec![],
-                    intent: None,
-                    boundary: None,
-                    redirects: vec![],
-                    source: Some("migrations/001.sql:1".into()),
-                })
-                .collect(),
-            ..Default::default()
-        }
-    }
-
-    #[tokio::test]
-    async fn schema_gate_blocks_duplicate_table_on_write() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["users"]));
-
-        // Attempt to write a SQL file that creates `users` again.
-        let result = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/002.sql",
-                    "content": "CREATE TABLE users (id INT);\n"
-                }),
-            )
-            .await;
-        assert!(result.is_error());
-        match result {
-            ToolOutput::Error { message } => {
-                assert!(
-                    message.contains("Table `users` already exists"),
-                    "{message}"
-                );
-                assert!(message.contains("ALTER"), "{message}");
-                // Gate v2: the conflict cites the canonical address and the
-                // existing columns, so the model can decide without a read.
-                assert!(message.contains("store://sql/default/users"), "{message}");
-                assert!(message.contains("columns: id INT"), "{message}");
-            }
-            _ => panic!("expected error"),
-        }
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_allows_new_table_on_write() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_ok_{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["users"]));
-
-        // Write a SQL file with a genuinely new, dissimilar table.
-        let result = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/003.sql",
-                    "content": "CREATE TABLE orders (id INT);\n"
-                }),
-            )
-            .await;
-        assert!(!result.is_error(), "new table should pass the gate");
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_blocks_normalized_duplicate_and_column_dup() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_norm_{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["user_profiles"]));
-
-        // `UserProfile` is the same relation as `user_profiles` after
-        // normalization + plural folding.
-        let camel = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/004.sql",
-                    "content": "CREATE TABLE UserProfile (id INT);\n"
-                }),
-            )
-            .await;
-        assert!(camel.is_error(), "normalized duplicate must be blocked");
-
-        // Adding a column that already exists (id) is column-level drift.
-        let column = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/005.sql",
-                    "content": "ALTER TABLE user_profiles ADD COLUMN Id BIGINT;\n"
-                }),
-            )
-            .await;
-        assert!(column.is_error(), "duplicate column must be blocked");
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_challenge_passes_with_declared_intent_and_records_it() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_intent_{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["payments"]));
-
-        // `payment_records` is not a name duplicate, but it resembles
-        // `payments` — withheld once with the evidence.
-        let call = serde_json::json!({
-            "path": "migrations/006.sql",
-            "content": "CREATE TABLE payment_records (id INT);\n"
-        });
-        let challenged = reg.execute("write_file", &call).await;
-        match &challenged {
-            ToolOutput::Error { message } => {
-                assert!(message.contains("write withheld"), "{message}");
-                assert!(message.contains("storage_intent"), "{message}");
-            }
-            _ => panic!("expected the similarity challenge"),
-        }
-
-        // Retrying with a declared intent passes, lands the file, and
-        // records the sentence in stella.storage.toml (origin `declared`).
-        let mut with_intent = call.clone();
-        with_intent["storage_intent"] = serde_json::json!(
-            "Immutable ledger of imported legacy charges; payments holds live charges."
-        );
-        let passed = reg.execute("write_file", &with_intent).await;
-        assert!(!passed.is_error(), "declared intent must pass: {passed:?}");
-        let manifest = std::fs::read_to_string(dir.join("stella.storage.toml")).unwrap();
-        assert!(
-            manifest.contains("sql/default/payment_records"),
-            "{manifest}"
-        );
-        assert!(manifest.contains("Immutable ledger"), "{manifest}");
-        assert!(manifest.contains("origin = \"declared\""), "{manifest}");
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_exempts_rewriting_a_files_own_objects() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_own_{}", std::process::id()));
-        std::fs::create_dir_all(dir.join("migrations")).unwrap();
-        std::fs::write(
-            dir.join("migrations/001.sql"),
-            "CREATE TABLE users (id INT);\n",
-        )
-        .unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["users"]));
-
-        // Rewriting the file that defines `users` is an in-place change to
-        // the existing object, not a duplicate creation.
-        let result = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/001.sql",
-                    "content": "CREATE TABLE users (id INT, email TEXT);\n"
-                }),
-            )
-            .await;
-        assert!(!result.is_error(), "same-file rewrite must pass the gate");
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_grows_the_index_after_a_successful_write() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_grow_{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-
-        // Empty index: the first migration passes and registers `users`.
-        let first = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/001.sql",
-                    "content": "CREATE TABLE users (id INT);\n"
-                }),
-            )
-            .await;
-        assert!(!first.is_error());
-
-        // A second file re-creating `users` now conflicts — caught by the
-        // in-session index growth, no graph re-index needed.
-        let second = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "migrations/002.sql",
-                    "content": "CREATE TABLE users (id INT);\n"
-                }),
-            )
-            .await;
-        assert!(second.is_error(), "in-session duplicate must be caught");
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_checks_edit_file_new_string() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_edit_{}", std::process::id()));
-        std::fs::create_dir_all(dir.join("migrations")).unwrap();
-        std::fs::write(dir.join("migrations/002.sql"), "-- add payments\n").unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["users"]));
-
-        // An edit whose replacement introduces a duplicate CREATE is gated
-        // exactly like a write.
-        let result = reg
-            .execute(
-                "edit_file",
-                &serde_json::json!({
-                    "path": "migrations/002.sql",
-                    "old_string": "-- add payments",
-                    "new_string": "CREATE TABLE users (id INT);"
-                }),
-            )
-            .await;
-        assert!(
-            result.is_error(),
-            "edit introducing a duplicate must be gated"
-        );
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    #[tokio::test]
-    async fn schema_gate_allows_non_sql_write() {
-        let dir = std::env::temp_dir().join(format!("stella_gate_nosql_{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.clone(), None);
-        reg.update_storage_index(seeded_snapshot(&["users"]));
-
-        // Writing a Rust file should never trigger the schema gate.
-        let result = reg
-            .execute(
-                "write_file",
-                &serde_json::json!({
-                    "path": "src/main.rs",
-                    "content": "fn main() {}\n"
-                }),
-            )
-            .await;
-        assert!(!result.is_error());
-        std::fs::remove_dir_all(&dir).ok();
-    }
-
-    // file-touch telemetry
-
-    /// Fresh registry over a fresh tempdir, no optional backends.
-    fn telemetry_fixture() -> (tempfile::TempDir, ToolRegistry) {
-        let dir = tempfile::tempdir().unwrap();
-        let reg = ToolRegistry::with_issue_backend(dir.path().to_path_buf(), None);
-        (dir, reg)
-    }
-
-    async fn exec_ok(reg: &ToolRegistry, name: &str, input: serde_json::Value) {
-        let out = reg.execute(name, &input).await;
-        assert!(!out.is_error(), "{name} {input} failed: {out:?}");
-    }
-
-    #[tokio::test]
-    async fn touch_read_only_file_records_one_r_with_zero_deltas() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("notes.txt"), "one\ntwo\n").unwrap();
-
-        exec_ok(&reg, "read_file", serde_json::json!({"path": "notes.txt"})).await;
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        assert_eq!(payload.files_touched.len(), 1);
-        let record = &payload.files_touched[0];
-        assert_eq!(record.path, "notes.txt");
-        assert_eq!(record.crud_events, vec![FileOp::Read]);
-        assert_eq!((record.lines_added, record.lines_removed), (0, 0));
-        assert_eq!(record.events.len(), 1);
-        assert_eq!(
-            (record.events[0].lines_added, record.events[0].lines_removed),
-            (0, 0)
-        );
-    }
-
-    #[tokio::test]
-    async fn touch_multiple_reads_are_one_record_with_multiple_events() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("notes.txt"), "one\ntwo\n").unwrap();
-
-        for _ in 0..3 {
-            exec_ok(&reg, "read_file", serde_json::json!({"path": "notes.txt"})).await;
-        }
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        assert_eq!(payload.files_touched.len(), 1, "one record per file");
-        let record = &payload.files_touched[0];
-        assert_eq!(record.crud_events, vec![FileOp::Read], "crud deduplicated");
-        assert_eq!(record.events.len(), 3, "audit log never deduplicated");
-    }
-
-    #[tokio::test]
-    async fn touch_create_counts_the_new_files_full_line_count() {
-        let (_dir, reg) = telemetry_fixture();
-
-        exec_ok(
-            &reg,
-            "write_file",
-            serde_json::json!({"path": "src/new.rs", "content": "a\nb\nc\n"}),
-        )
-        .await;
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        let record = &payload.files_touched[0];
-        assert_eq!(record.path, "src/new.rs");
-        assert_eq!(record.crud_events, vec![FileOp::Create]);
-        assert_eq!((record.lines_added, record.lines_removed), (3, 0));
-    }
-
-    #[tokio::test]
-    async fn touch_update_counts_the_line_diff() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("code.rs"), "a\nb\nc\n").unwrap();
-
-        // edit_file: "b" becomes two lines — +2 −1.
-        exec_ok(
-            &reg,
-            "edit_file",
-            serde_json::json!({"path": "code.rs", "old_string": "b", "new_string": "x\ny"}),
-        )
-        .await;
-        // write_file over an existing file is also an update — the diff of
-        // pre vs post, not the full content: "a\nx\ny\nc\n" → "a\nc\n" is −2.
-        exec_ok(
-            &reg,
-            "write_file",
-            serde_json::json!({"path": "code.rs", "content": "a\nc\n"}),
-        )
-        .await;
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        let record = &payload.files_touched[0];
-        assert_eq!(record.crud_events, vec![FileOp::Update]);
-        assert_eq!(record.events.len(), 2);
-        assert_eq!(
-            (record.events[0].lines_added, record.events[0].lines_removed),
-            (2, 1)
-        );
-        assert_eq!(
-            (record.events[1].lines_added, record.events[1].lines_removed),
-            (0, 2)
-        );
-        assert_eq!((record.lines_added, record.lines_removed), (2, 3));
-    }
-
-    #[tokio::test]
-    async fn touch_delete_counts_the_full_pre_deletion_line_count() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("doomed.txt"), "1\n2\n3\n4\n5\n").unwrap();
-
-        exec_ok(
-            &reg,
-            "delete_file",
-            serde_json::json!({"path": "doomed.txt"}),
-        )
-        .await;
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        let record = &payload.files_touched[0];
-        assert_eq!(record.crud_events, vec![FileOp::Delete]);
-        assert_eq!((record.lines_added, record.lines_removed), (0, 5));
-    }
-
-    #[tokio::test]
-    async fn touch_read_update_delete_keeps_order_and_dedupes_crud() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("f.rs"), "a\nb\nc\n").unwrap();
-
-        exec_ok(
-            &reg,
-            "read_file",
-            serde_json::json!({"path": "f.rs", "reason": "inspect before edit"}),
-        )
-        .await;
-        exec_ok(
-            &reg,
-            "edit_file",
-            serde_json::json!({
-                "path": "f.rs", "old_string": "b", "new_string": "x\ny",
-                "reason": "split b into x and y"
-            }),
-        )
-        .await;
-        exec_ok(
-            &reg,
-            "delete_file",
-            serde_json::json!({"path": "f.rs", "reason": "no longer needed"}),
-        )
-        .await;
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        assert_eq!(payload.files_touched.len(), 1);
-        let record = &payload.files_touched[0];
-        assert_eq!(
-            record.crud_events,
-            vec![FileOp::Read, FileOp::Update, FileOp::Delete],
-            "deduplicated, first-occurrence order"
-        );
-        let ops: Vec<FileOp> = record.events.iter().map(|e| e.event).collect();
-        assert_eq!(ops, vec![FileOp::Read, FileOp::Update, FileOp::Delete]);
-        let reasons: Vec<&str> = record.events.iter().map(|e| e.reason.as_str()).collect();
-        assert_eq!(
-            reasons,
-            vec![
-                "inspect before edit",
-                "split b into x and y",
-                "no longer needed"
-            ]
-        );
-        // R: 0/0. U: +2 −1. D: −4 (the post-edit file "a\nx\ny\nc\n").
-        assert_eq!((record.lines_added, record.lines_removed), (2, 5));
-    }
-
-    #[tokio::test]
-    async fn touch_equivalent_path_spellings_aggregate_into_one_record() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::create_dir_all(dir.path().join("src")).unwrap();
-        std::fs::write(dir.path().join("src/main.rs"), "fn main() {}\n").unwrap();
-
-        for spelling in [
-            "src/main.rs",
-            "src/./main.rs",
-            "src/../src/main.rs",
-            "./src/main.rs",
-        ] {
-            exec_ok(&reg, "read_file", serde_json::json!({"path": spelling})).await;
-        }
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        assert_eq!(
-            payload.files_touched.len(),
-            1,
-            "equivalent spellings must aggregate: {payload:?}"
-        );
-        let record = &payload.files_touched[0];
-        assert_eq!(record.path, "src/main.rs");
-        assert_eq!(record.events.len(), 4);
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 8)]
-    async fn touch_concurrent_tools_lose_no_events_and_keep_totals_exact() {
-        let (dir, reg) = telemetry_fixture();
-        for i in 0..4 {
-            std::fs::write(dir.path().join(format!("read{i}.txt")), "x\n").unwrap();
-        }
-        let reg = std::sync::Arc::new(reg);
-
-        let mut handles = Vec::new();
-        // 32 concurrent reads across 4 files (read-only tools are the ones
-        // the engine actually parallelizes)...
-        for i in 0..32 {
-            let reg = reg.clone();
-            handles.push(tokio::spawn(async move {
-                let input = serde_json::json!({"path": format!("read{}.txt", i % 4)});
-                let out = reg.execute("read_file", &input).await;
-                assert!(!out.is_error(), "{out:?}");
-            }));
-        }
-        // ...plus 8 concurrent creates of distinct files.
-        for i in 0..8 {
-            let reg = reg.clone();
-            handles.push(tokio::spawn(async move {
-                let input = serde_json::json!({
-                    "path": format!("new{i}.txt"),
-                    "content": "l1\nl2\n"
-                });
-                let out = reg.execute("write_file", &input).await;
-                assert!(!out.is_error(), "{out:?}");
-            }));
-        }
-        for handle in handles {
-            handle.await.unwrap();
-        }
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        assert_eq!(payload.files_touched.len(), 12, "4 read + 8 created files");
-        let mut read_events = 0;
-        let mut created_lines = 0;
-        for record in &payload.files_touched {
-            if record.path.starts_with("read") {
-                assert_eq!(record.crud_events, vec![FileOp::Read]);
-                read_events += record.events.len();
-            } else {
-                assert_eq!(record.crud_events, vec![FileOp::Create]);
-                assert_eq!(record.events.len(), 1);
-                created_lines += record.lines_added;
-            }
-        }
-        assert_eq!(read_events, 32, "no read event may be lost");
-        assert_eq!(created_lines, 16, "8 files × 2 lines");
-    }
-
-    #[tokio::test]
-    async fn touch_failed_operations_record_nothing() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("f.rs"), "a\n").unwrap();
-
-        // Missing file, escaping path, and a failed edit precondition.
-        for (tool, input) in [
-            ("read_file", serde_json::json!({"path": "ghost.txt"})),
-            ("read_file", serde_json::json!({"path": "../../etc/passwd"})),
-            ("delete_file", serde_json::json!({"path": "ghost.txt"})),
-            (
-                "edit_file",
-                serde_json::json!({"path": "f.rs", "old_string": "nope", "new_string": "x"}),
-            ),
-        ] {
-            let out = reg.execute(tool, &input).await;
-            assert!(out.is_error(), "{tool} {input} should fail");
-        }
-
-        assert!(
-            reg.file_touch_telemetry().files_touched.is_empty(),
-            "failed operations must not create CRUD events"
-        );
-    }
-
-    #[tokio::test]
-    async fn touch_reason_defaults_when_omitted() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("f.rs"), "a\n").unwrap();
-
-        exec_ok(&reg, "read_file", serde_json::json!({"path": "f.rs"})).await;
-
-        let payload = reg.file_touch_telemetry();
-        payload.validate().unwrap();
-        let event = &payload.files_touched[0].events[0];
-        assert!(
-            !event.reason.is_empty(),
-            "schema requires a non-empty reason"
-        );
-    }
-
-    #[tokio::test]
-    async fn files_touched_compact_view_still_uses_crud_display_order() {
-        // The legacy (path, "RU") view feeds the TUI panel and episodic
-        // memory: letters stay in C,R,U,D display order even though the
-        // telemetry payload orders by first occurrence.
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("f.rs"), "a\nb\n").unwrap();
-
-        exec_ok(
-            &reg,
-            "edit_file",
-            serde_json::json!({"path": "f.rs", "old_string": "a", "new_string": "z"}),
-        )
-        .await;
-        exec_ok(&reg, "read_file", serde_json::json!({"path": "f.rs"})).await;
-
-        assert_eq!(
-            reg.files_touched(),
-            vec![("f.rs".to_string(), "RU".to_string())]
-        );
-        assert_eq!(
-            reg.file_touch_telemetry().files_touched[0].crud_events,
-            vec![FileOp::Update, FileOp::Read],
-            "payload keeps first-occurrence order"
-        );
-    }
-
-    // extension hook bus integration
-
-    use std::sync::Arc as StdArc;
-    use std::sync::Mutex as StdMutex;
-    use stella_core::bus::{HookBus, HookDecision, HookEvent, names as hook_names};
-
-    /// Capture every event a `"*"` observer sees on `bus`.
-    fn record_bus_events(bus: &HookBus) -> StdArc<StdMutex<Vec<HookEvent>>> {
-        let seen = StdArc::new(StdMutex::new(Vec::new()));
-        let sink = seen.clone();
-        bus.on("*", move |event| {
-            sink.lock().unwrap().push(event.clone());
-            Ok(())
-        })
-        .detach();
-        seen
-    }
-
-    fn event_names(seen: &StdArc<StdMutex<Vec<HookEvent>>>) -> Vec<String> {
-        seen.lock()
-            .unwrap()
-            .iter()
-            .map(|e| e.name.clone())
-            .collect()
-    }
-
-    #[tokio::test]
-    async fn a_deny_policy_blocks_the_tool_and_leaves_no_touch() {
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("f.rs"), "keep me\n").unwrap();
-        let bus = HookBus::new("sess");
-        bus.on_blocking(hook_names::TOOL_CALL_REQUESTED, |_| HookDecision::Deny {
-            reason: "workspace is read-only".into(),
-        })
-        .detach();
-        reg.attach_bus(bus);
-
-        // The delete is refused by the policy; the file survives and the
-        // ledger records nothing.
-        let out = reg
-            .execute("delete_file", &serde_json::json!({"path": "f.rs"}))
-            .await;
-        assert!(out.is_error(), "deny must surface as a tool error");
-        match out {
-            ToolOutput::Error { message } => assert!(message.contains("read-only"), "{message}"),
-            _ => unreachable!(),
-        }
-        assert!(
-            dir.path().join("f.rs").exists(),
-            "denied delete must not run"
-        );
-        assert!(
-            reg.file_touch_telemetry().files_touched.is_empty(),
-            "a blocked op records no file touch"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_successful_write_emits_the_file_and_files_touched_events() {
-        let (_dir, reg) = telemetry_fixture();
-        let bus = HookBus::new("sess");
-        let seen = record_bus_events(&bus);
-        reg.attach_bus(bus);
-
-        exec_ok(
-            &reg,
-            "write_file",
-            serde_json::json!({"path": "src/new.rs", "content": "a\nb\nc\n", "reason": "scaffold"}),
-        )
-        .await;
-
-        let names = event_names(&seen);
-        // The OBSERVABLE lifecycle for a create. The raw `tool.call.requested`
-        // / `file.created` blocking events are delivered only to blocking
-        // (privileged) handlers, never to observers — observers see their
-        // safe outcomes (`policy.*`), the sanitized `tool.call.started`, the
-        // completion, and the `file.created` FACT + aggregate update.
-        for expected in [
-            hook_names::POLICY_ALLOWED,
-            hook_names::TOOL_CALL_STARTED,
-            hook_names::TOOL_CALL_COMPLETED,
-            hook_names::FILE_CREATED,
-            hook_names::FILES_TOUCHED_UPDATED,
-        ] {
-            assert!(
-                names.iter().any(|n| n == expected),
-                "missing {expected} in {names:?}"
-            );
-        }
-        // The raw blocking event never reaches an observer.
-        assert!(
-            !names.iter().any(|n| n == hook_names::TOOL_CALL_REQUESTED),
-            "raw tool.call.requested must not reach observers: {names:?}"
-        );
-        // The file.created fact carries the line delta but never the content.
-        let events = seen.lock().unwrap();
-        let created = events
-            .iter()
-            .find(|e| e.name == hook_names::FILE_CREATED)
-            .unwrap();
-        assert_eq!(created.payload["path"], "src/new.rs");
-        assert_eq!(created.payload["lines_added"], 3);
-        let started = events
-            .iter()
-            .find(|e| e.name == hook_names::TOOL_CALL_STARTED)
-            .unwrap();
-        assert!(
-            started.payload["input"]["content"]
-                .as_str()
-                .unwrap()
-                .starts_with("<omitted:"),
-            "tool.call.started must not leak file content"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_modify_policy_rewrites_the_input_the_tool_runs() {
-        let (dir, reg) = telemetry_fixture();
-        let bus = HookBus::new("sess");
-        bus.on_blocking(hook_names::TOOL_CALL_REQUESTED, |event| {
-            // Redirect any write to a quarantine path.
-            let mut input = event.payload["input"].clone();
-            input["path"] = serde_json::Value::String("quarantine.txt".into());
-            let mut payload = event.payload.clone();
-            payload["input"] = input;
-            HookDecision::Modify { payload }
-        })
-        .detach();
-        reg.attach_bus(bus);
-
-        exec_ok(
-            &reg,
-            "write_file",
-            serde_json::json!({"path": "original.txt", "content": "x\n"}),
-        )
-        .await;
-
-        assert!(
-            dir.path().join("quarantine.txt").exists(),
-            "the modified path is what actually got written"
-        );
-        assert!(!dir.path().join("original.txt").exists());
-        assert_eq!(
-            reg.file_touch_telemetry().files_touched[0].path,
-            "quarantine.txt",
-            "the ledger records the path the tool actually touched"
-        );
-    }
-
-    #[tokio::test]
-    async fn cite_memory_dispatches_through_the_registry_and_drains_once() {
-        let (_dir, reg) = telemetry_fixture();
-        exec_ok(
-            &reg,
-            "cite_memory",
-            serde_json::json!({
-                "memory_id": "nod_0123456789abcdef01234567",
-                "useful_score": 5,
-                "truthful": true,
-                "remark": "named the failing module outright",
-            }),
-        )
-        .await;
-        let drained = reg.take_memory_citations();
-        assert_eq!(drained.len(), 1);
-        assert_eq!(drained[0].memory_id, "nod_0123456789abcdef01234567");
-        assert!(
-            reg.take_memory_citations().is_empty(),
-            "a drained citation must never persist under a second execution"
-        );
-    }
-
-    #[tokio::test]
-    async fn a_denied_command_never_runs_and_a_bus_less_registry_is_unchanged() {
-        // Command guards apply to `bash`, which ships registered.
-        let dir = tempfile::tempdir().unwrap();
-        let reg = ToolRegistry::with_backends_and_options(
-            dir.path().to_path_buf(),
-            None,
-            None,
-            RegistryOptions::default(),
-        );
-        let bus = HookBus::new("sess");
-        bus.on_blocking(hook_names::COMMAND_STARTED, |_| HookDecision::Deny {
-            reason: "no shell".into(),
-        })
-        .detach();
-        reg.attach_bus(bus);
-        let out = reg
-            .execute("bash", &serde_json::json!({"command": "echo hi"}))
-            .await;
-        assert!(out.is_error());
-
-        // A registry with no bus attached behaves exactly as before hooks.
-        let (dir2, plain) = telemetry_fixture();
-        std::fs::write(dir2.path().join("f.txt"), "hi\n").unwrap();
-        exec_ok(&plain, "read_file", serde_json::json!({"path": "f.txt"})).await;
-        assert_eq!(plain.file_touch_telemetry().files_touched.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn the_command_chain_gates_run_script_with_its_resolved_command() {
-        // `run_script` composes its command from the scripts index, so the
-        // same `command.started` policy that fences `bash` must fence it —
-        // with the resolved command line, not the script name.
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("Makefile"), "greet:\n\t@echo hi\n").unwrap();
-        let bus = HookBus::new("sess");
-        let denied_command = StdArc::new(StdMutex::new(String::new()));
-        let sink = denied_command.clone();
-        bus.on_blocking(hook_names::COMMAND_STARTED, move |event| {
-            *sink.lock().unwrap() = event.payload["command"].as_str().unwrap_or("").to_string();
-            HookDecision::Deny {
-                reason: "no shell".into(),
-            }
-        })
-        .detach();
-        reg.attach_bus(bus);
-
-        let out = reg
-            .execute("run_script", &serde_json::json!({"script": "make:greet"}))
-            .await;
-        assert!(
-            out.is_error(),
-            "denied run_script must not execute: {out:?}"
-        );
-        assert_eq!(
-            *denied_command.lock().unwrap(),
-            "make greet",
-            "the chain must see the index-resolved command line"
-        );
-    }
-
-    #[tokio::test]
-    async fn the_command_chain_gates_start_process_with_its_joined_argv() {
-        // `start_process` sits in the DEFAULT surface while `bash` is
-        // opt-in, and argv[0] may itself be a shell — so the same
-        // `command.started` policy that fences `bash` must fence the argv
-        // spawn, seeing the joined argv, before anything runs.
-        let (_dir, reg) = telemetry_fixture();
-        let bus = HookBus::new("sess");
-        let denied_command = StdArc::new(StdMutex::new(String::new()));
-        let sink = denied_command.clone();
-        bus.on_blocking(hook_names::COMMAND_STARTED, move |event| {
-            *sink.lock().unwrap() = event.payload["command"].as_str().unwrap_or("").to_string();
-            HookDecision::Deny {
-                reason: "no shell".into(),
-            }
-        })
-        .detach();
-        reg.attach_bus(bus);
-
-        let out = reg
-            .execute(
-                "start_process",
-                &serde_json::json!({"argv": ["bash", "-c", "echo hi"]}),
-            )
-            .await;
-        assert!(
-            out.is_error(),
-            "denied start_process must not spawn: {out:?}"
-        );
-        assert_eq!(
-            *denied_command.lock().unwrap(),
-            "bash -c echo hi",
-            "the chain must see the joined argv as the command line"
-        );
-        // The denial fired before the spawn: no handle may exist.
-        let read = reg
-            .execute("read_output", &serde_json::json!({"handle": "proc-1"}))
-            .await;
-        assert!(
-            read.is_error(),
-            "no process may exist after a denied spawn: {read:?}"
-        );
-    }
-
-    /// The hole the `start_process` gate left open. Gating only the spawn
-    /// fences `["bash", "-c", "cat"]` once, on that one argv; every command
-    /// pushed into the live shell afterwards used to execute with no policy
-    /// consultation and no `command.*` audit record, so a session could run
-    /// arbitrary shell through a channel that is on by default while `bash`
-    /// is opt-in.
-    #[tokio::test]
-    async fn the_command_chain_gates_the_text_written_into_a_live_interpreter() {
-        let (_dir, reg) = telemetry_fixture();
-        let bus = HookBus::new("sess");
-        let seen = StdArc::new(StdMutex::new(Vec::<String>::new()));
-        let sink = seen.clone();
-        // Allow the spawn, deny what gets written into it — the operator
-        // posture that permits a REPL but not arbitrary commands inside it.
-        bus.on_blocking(hook_names::COMMAND_STARTED, move |event| {
-            let command = event.payload["command"].as_str().unwrap_or("").to_string();
-            sink.lock().unwrap().push(command.clone());
-            if command.contains("rm -rf") {
-                HookDecision::Deny {
-                    reason: "no destructive commands".into(),
-                }
-            } else {
-                HookDecision::Allow
-            }
-        })
-        .detach();
-        reg.attach_bus(bus);
-
-        let spawned = reg
-            .execute("start_process", &serde_json::json!({"argv": ["cat"]}))
-            .await;
-        assert!(!spawned.is_error(), "the spawn is allowed: {spawned:?}");
-
-        let out = reg
-            .execute(
-                "send_stdin",
-                &serde_json::json!({"handle": "proc-1", "text": "rm -rf /\n"}),
-            )
-            .await;
-        assert!(
-            out.is_error(),
-            "writing a denied command into a live interpreter must not go \
-             through: {out:?}"
-        );
-        let seen = seen.lock().unwrap().clone();
-        assert!(
-            seen.iter().any(|c| c.contains("rm -rf")),
-            "the chain must be shown the stdin text, not just the spawn argv; \
-             it saw {seen:?}"
-        );
-    }
-
-    /// A `command.started` policy that denies shell execution, capturing
-    /// every command line the chain was shown — the posture an extension
-    /// uses to fence command execution.
-    fn deny_shell(reg: &ToolRegistry) -> StdArc<StdMutex<Vec<String>>> {
-        let seen = StdArc::new(StdMutex::new(Vec::new()));
-        let sink = seen.clone();
-        let bus = HookBus::new("sess");
-        bus.on_blocking(hook_names::COMMAND_STARTED, move |event| {
-            sink.lock()
-                .unwrap()
-                .push(event.payload["command"].as_str().unwrap_or("").to_string());
-            HookDecision::Deny {
-                reason: "no shell".into(),
-            }
-        })
-        .detach();
-        reg.attach_bus(bus);
-        seen
-    }
-
-    #[tokio::test]
-    async fn the_command_chain_gates_the_explicit_build_test_and_verify_commands() {
-        // `build_project`, `run_tests`, and `verify_done` all reach
-        // `bash -c` from the DEFAULT surface while `bash` is opt-in, so the
-        // same `command.started` policy that fences `bash` must fence
-        // them — seeing the model's own command text, before anything runs.
-        let (dir, reg) = telemetry_fixture();
-        let commands = deny_shell(&reg);
-
-        for (tool, input) in [
-            (
-                "build_project",
-                serde_json::json!({"command": "touch build-ran.txt"}),
-            ),
-            (
-                "run_tests",
-                serde_json::json!({"command": "touch tests-ran.txt"}),
-            ),
-            (
-                "verify_done",
-                serde_json::json!({
-                    "test_cmd": "touch verify-ran.txt",
-                    "test_files": ["a.rs"],
-                }),
-            ),
-        ] {
-            let out = reg.execute(tool, &input).await;
-            assert!(out.is_error(), "denied {tool} must not execute: {out:?}");
-        }
-
-        assert_eq!(
-            *commands.lock().unwrap(),
-            vec![
-                "touch build-ran.txt",
-                "touch tests-ran.txt",
-                "touch verify-ran.txt",
-            ],
-            "the chain must see each tool's own command line"
-        );
-        // Each denial fired before the shell ran: no marker file exists.
-        for marker in ["build-ran.txt", "tests-ran.txt", "verify-ran.txt"] {
-            assert!(
-                !dir.path().join(marker).exists(),
-                "`{marker}` proves a denied command still ran"
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn the_command_chain_gates_the_index_composed_build_and_test_commands() {
-        // The common case is the model omitting `command` entirely and the
-        // tool composing one from the scripts index — that path must be
-        // fenced with the composed line too, or `{}` reopens the bypass.
-        let (dir, reg) = telemetry_fixture();
-        std::fs::write(dir.path().join("Cargo.toml"), "[package]\nname = \"x\"\n").unwrap();
-        let commands = deny_shell(&reg);
-
-        let out = reg.execute("build_project", &serde_json::json!({})).await;
-        assert!(out.is_error(), "denied build_project must not run: {out:?}");
-        let out = reg
-            .execute(
-                "run_tests",
-                &serde_json::json!({"kind": "unit", "filter": "my_test"}),
-            )
-            .await;
-        assert!(out.is_error(), "denied run_tests must not run: {out:?}");
-
-        assert_eq!(
-            *commands.lock().unwrap(),
-            vec![
-                "cargo build --workspace",
-                // The filter is shell-quoted per token, so the gated line is
-                // still byte-identical to what would have executed — which
-                // is the property this test exists to protect.
-                "cargo test --workspace --lib --bins 'my_test'",
-            ],
-            "the chain must see the index-composed command line"
-        );
-    }
-}
+mod tests;
